refactor(synth): extract shared synthesis-retriever module (#33)
Move the retrieval-enriched synthesis helpers out of analyze-sessions.ts
into scripts/knowledge-graph/synthesis-retriever.ts so both entry points
(analyze-sessions and the npx CLI) can share one implementation instead of
re-implementing the build strategy + per-candidate retrieve loop.

Exports buildSynthesisRetriever (3 strategies), retrieveContextForCandidate
(query-build + retrieve + try/catch->undefined fallback), and a single
RETRIEVAL_TOP_K constant. analyze-sessions now consumes both; cli.ts is
rewired in a follow-up commit.
